Manage front-end dependencies with gulp
This change makes it easier to manage front end dependencies.
A combinaison of Gulp and Bower is used to fetch and install all
the required files.

This change made it possible to remove the Braincrafted and
Assitic bundles.  It was also possible to remove many dependencies
specified in composer.json as they are now defined in bower.json.